CHG: move some utils into LLMProxy
