+++ tests/llm/stream_test.py
@@ -8,7 +8,7 @@ from litellm import ModelResponse
 from promptmodel.llms.llm import LLM
 from promptmodel.llms.llm_proxy import LLMProxy
 from promptmodel.types.response import LLMResponse, LLMStreamResponse
-from promptmodel.utils.prompt_util import run_async_in_sync
+from promptmodel.utils.async_util import run_async_in_sync
 
 
 def test_stream(mocker):
